redesign model

+++ database/migrations/2023_06_29_081707_create_bahan_baku_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('bahan_baku', function (Blueprint $table) {
             $table->uuid('id')->primary();
             $table->string('nama');
-            $table->float('stok');
+            $table->integer('stok');
             $table->string('satuan');
             $table->softDeletes();
             $table->timestamps();
